Extract UAT actions behaviour
