+++ .rubocop.yml
@@ -100,6 +100,8 @@ Metrics/BlockLength:
     - redefine_aasm
     - model
     - step
+    - class_methods
+    - included
 # We have a few custom find_by methods which we maintain.
 # At least for now. find_by_id_or_name is a bit clunky.
 Rails/DynamicFindBy:
